requirements.txt
Django==1.9.7
djangorestframework==3.3.3
requests==2.10.0
wheel==0.24.0
